requirements.txt
absl-py==1.0.0
deepillusion==0.3.3
hydra-core==1.1.1
joypy==0.2.6
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
scipy==1.8.0
simple-gpu-scheduler==0.1.4
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
XlsxWriter==3.0.3